Params from wildcards trump params from query string
If a URL has {foo} in it and the request also contains ?foo=bar, make
the value that matched the wildcard take precedence over what was passed
in the query string.

Simply switching the order in which the internal URL is constructed is
actually sufficient, but the fact that you can have multiple instances
of the same value still leaves me nervous. If you put something as a
wildcard in the URL, you are expecting one instance of it. Anything on
top of that is killed with fire.