Allow Marshaled to respond with arbitrary streams.
If the return type of a Marshaled handler interface implements the
io.Reader interface the stream will be written directly to the
requestor. A 'Content-Type' header is required to be specified in the
response headers and the 'Accept' header for these particular requests
can be anything.

Additionally, if the return type of the Marshaled handler implements the
io.Closer interface the stream will be automatically closed after it is
flushed to the requestor.

Logging of non-JSON responses bodies has been redacted to prevent
flushing of binary responses to logs.